requirements-dev.txt
-r requirements.txt
pre-commit>=3.4.0
ansible-lint>=6.17.2
yamllint>=1.32.0
ipaddress>=1.0.23
pyyaml>=6.0.1
pytest>=7.4.2
pytest-cov>=4.1.0